Code managment and adding required files for execution.
