+++ .gitignore
@@ -1,3 +1,1 @@
-*.mod
-*.sum
 *.sh
