+++ go.mod
@@ -1,0 +1,10 @@
+module web-chat-go
+
+go 1.18
+
+require (
+	github.com/google/uuid v1.3.0
+	github.com/gorilla/websocket v1.5.0
+	github.com/lib/pq v1.10.6
+	golang.org/x/crypto v0.0.0-20220622213112-05595931fe9d
+)
